starina/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
testfixtures==6.14.2
